move mda to gui folder
Required some extra imports in explore-sample. Will remove in
a subsequent commit that decouples them.
